pyproject.toml
[build-system]
requires = ["hatchling>=1.5.0", "hatch-nodejs-version>=0.3.2", "hatch-jupyter-builder>=0.9.1"]
build-backend = "hatchling.build"

[project]
name = "marimo-jupyter-scheduler"
dynamic = ["version", "description", "authors", "urls"]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
keywords = ["jupyter", "jupyterlab", "marimo", "scheduler", "jupyterhub"]
classifiers = [
    "Framework :: Jupyter",
    "Framework :: Jupyter :: JupyterLab",
    "Framework :: Jupyter :: JupyterLab :: 4",
    "Framework :: Jupyter :: JupyterLab :: Extensions",
    "Framework :: Jupyter :: JupyterLab :: Extensions :: Prebuilt",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "jupyter_scheduler>=2.5.0,<3",
    "jupyterlab>=4.0.0,<5",
    "jupyter_server>=2.0.1,<3",
    "marimo>=0.9.0",
    "pyyaml>=6.0",
    "watchdog>=3.0.0",
    "sqlalchemy>=2.0.0",
    "apscheduler>=3.10.0",
    "traitlets>=5.0",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9.0"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-tornasync",
    "coverage",
    "mypy",
    "ruff",
]


[tool.hatch.version]
source = "nodejs"

[tool.hatch.metadata.hooks.nodejs]
fields = ["description", "authors", "urls"]

[tool.hatch.build.targets.sdist]
artifacts = ["marimo_jupyter_scheduler/labextension"]
exclude = [".github", "binder"]

[tool.hatch.build.targets.wheel]
artifacts = ["marimo_jupyter_scheduler/labextension"]

[tool.hatch.build.targets.wheel.shared-data]
"marimo_jupyter_scheduler/labextension" = "share/jupyter/labextensions/marimo-jupyter-scheduler"
"install.json" = "share/jupyter/labextensions/marimo-jupyter-scheduler/install.json"
"jupyter-config/server-config" = "etc/jupyter/jupyter_server_config.d"

[tool.hatch.build.hooks.version]
path = "marimo_jupyter_scheduler/_version.py"

[tool.hatch.build.hooks.jupyter-builder]
dependencies = ["hatch-jupyter-builder>=0.9.1"]
build-function = "hatch_jupyter_builder.npm_builder"
ensured-targets = [
    "marimo_jupyter_scheduler/labextension/static/style.js",
    "marimo_jupyter_scheduler/labextension/package.json",
]
skip-if-exists = ["marimo_jupyter_scheduler/labextension/static/style.js"]

[tool.hatch.build.hooks.jupyter-builder.build-kwargs]
build_cmd = "build:prod"
npm = ["jlpm"]

[tool.hatch.build.hooks.jupyter-builder.editable-build-kwargs]
build_cmd = "install:extension"
npm = ["jlpm"]
source_dir = "src"
build_dir = "marimo_jupyter_scheduler/labextension"

[tool.jupyter-releaser.options]
version-cmd = "hatch version"

[tool.jupyter-releaser.hooks]
before-bump-version = ["python -m pip install hatch hatch-nodejs-version", "jlpm"]
before-release = ["python -m pip install hatch hatch-nodejs-version", "jlpm", "jlpm run build:prod"]

[tool.check-wheel-contents]
ignore = ["W002"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I"]
